DEV: organise analysis method into subfolders + revision of the base classes
