+++ spectrochempy/analysis/_baseclass/_analysisbase.py
@@ -8,254 +8,30 @@
 This module implements the base abstract classes to define estimators such as PCA, ...
 """
 
-import inspect
 import logging
 import warnings
-from copy import copy
-from functools import partial, update_wrapper
 
 import matplotlib.pyplot as plt
 import numpy as np
 import traitlets as tr
 from sklearn import linear_model
 
-from spectrochempy.application.metaconfigurable import MetaConfigurable
-from spectrochempy.core import app, set_loglevel
+from spectrochempy.core import app
 from spectrochempy.core.dataset.baseobjects.ndarray import NDArray
-from spectrochempy.core.dataset.coordset import CoordSet
 from spectrochempy.core.dataset.nddataset import NDDataset
 from spectrochempy.extern.traittypes import Array
-from spectrochempy.utils import exceptions
-from spectrochempy.utils.constants import MASKED, NOMASK
-from spectrochempy.utils.decorators import deprecated, preserve_signature
+from spectrochempy.utils.baseconfigurable import BaseConfigurable
+from spectrochempy.utils.decorators import _wrap_ndarray_output_to_nddataset, deprecated
 from spectrochempy.utils.docstrings import _docstring
+from spectrochempy.utils.exceptions import NotFittedError
 from spectrochempy.utils.plots import NBlue, NGreen, NRed
 from spectrochempy.utils.traits import NDDatasetType
 
 
 # ======================================================================================
-# Exceptions for analysis models
-# ======================================================================================
-class NotFittedError(exceptions.SpectroChemPyError):
-    """
-    Exception raised when an analysis estimator is not fitted
-    but one use one of its method.
-
-    Parameters
-    ----------
-    attr : method, optional
-        The method from which the error was issued. In general it is determined
-        automatically.
-    """
-
-    def __init__(self, attr=None):
-        frame = inspect.currentframe().f_back
-        caller = frame.f_code.co_name if attr is None else attr
-        model = frame.f_locals["self"].name
-        message = (
-            f"To use `{caller}` ,  the method `fit` of model `{model}`"
-            f" should be executed first"
-        )
-        super().__init__(message)
-
-
-# ======================================================================================
-# A decorator to transform np.ndarray output from models to NDDataset
-# according to the X (default) and/or Y input
-# ======================================================================================
-class _set_output(object):
-    def __init__(
-        self,
-        method,
-        *args,
-        meta_from="_X",  # the attribute or tuple of attributes from which meta data are taken
-        units="keep",
-        title="keep",
-        typex=None,
-        typey=None,
-        typesingle=None,
-    ):
-        self.method = method
-        update_wrapper(self, method)
-        self.meta_from = meta_from
-        self.units = units
-        self.title = title
-        self.typex = typex
-        self.typey = typey
-        self.typesingle = typesingle
-
-    @preserve_signature
-    def __get__(self, obj, objtype):
-        """Support instance methods."""
-        newfunc = partial(self.__call__, obj)
-        update_wrapper(newfunc, self.method)
-        return newfunc
-
-    def __call__(self, obj, *args, **kwargs):
-
-        from spectrochempy.core.dataset.coord import Coord
-        from spectrochempy.core.dataset.nddataset import NDDataset
-
-        # HACK to be able to used deprecated alias of the method, without error
-        # because if not this modification obj appears two times
-        if args and type(args[0]) == type(obj):
-            args = args[1:]
-
-        # get the method output - one or two arrays depending on the method and *args
-        output = self.method(obj, *args, **kwargs)
-
-        # restore eventually masked rows and columns
-        axis = "both"
-        if self.typex is not None and self.typex != "features":
-            axis = 0
-        elif self.typey is not None:
-            axis = 1
-
-        # if a single array was returned...
-        if not isinstance(output, tuple):
-            # ... make a tuple of 1 array:
-            data_tuple = (output,)
-            # ... and a tuple of 1 from_meta element:
-            if not isinstance(self.meta_from, tuple):
-                meta_from_tuple = (self.meta_from,)
-            else:
-                # ensure that the first one
-                meta_from_tuple = (self.meta_from[0],)
-        else:
-            data_tuple = output
-            meta_from_tuple = self.meta_from
-
-        out = []
-        for data, meta_from in zip(data_tuple, meta_from_tuple):
-            X_transf = NDDataset(data)
-
-            # Now set the NDDataset attributes from the original X
-
-            # determine the input X dataset
-            X = getattr(obj, meta_from)
-
-            if self.units is not None:
-                if self.units == "keep":
-                    X_transf.units = X.units
-                else:
-                    X_transf.units = self.units
-            X_transf.name = f"{X.name}_{obj.name}.{self.method.__name__}"
-            X_transf.history = f"Created using method {obj.name}.{self.method.__name__}"
-            if self.title is not None:
-                if self.title == "keep":
-                    X_transf.title = X.title
-                else:
-                    X_transf.title = self.title
-            # make coordset
-            M, N = X.shape
-            if X_transf.shape == X.shape and self.typex is None and self.typey is None:
-                X_transf.set_coordset(y=X.coord(0), x=X.coord(1))
-            else:
-                if self.typey == "components":
-                    X_transf.set_coordset(
-                        y=Coord(
-                            None,
-                            labels=["#%d" % (i) for i in range(X_transf.shape[0])],
-                            title="components",
-                        ),
-                        x=X.coord(-1).copy() if X.coord(-1) is not None else None,
-                    )
-                if self.typex == "components":
-                    X_transf.set_coordset(
-                        y=X.coord(0).copy() if X.coord(0) is not None else None,
-                        # cannot use X.y in case of transposed X
-                        x=Coord(
-                            None,
-                            labels=["#%d" % (i) for i in range(X_transf.shape[-1])],
-                            title="components",
-                        ),
-                    )
-                if self.typex == "features":
-                    X_transf.set_coordset(
-                        y=Coord(
-                            None,
-                            labels=["#%d" % (i) for i in range(X_transf.shape[-1])],
-                            title="components",
-                        ),
-                        x=X.coord(1).copy() if X.coord(1) is not None else None,
-                    )
-                if self.typey == "features":
-                    X_transf.set_coordset(
-                        y=X.coord(1).copy() if X.coord(1) is not None else None,
-                        x=Coord(
-                            None,
-                            labels=["#%d" % (i) for i in range(X_transf.shape[-1])],
-                            title="components",
-                        ),
-                    )
-                if self.typesingle == "components":
-                    # occurs when the data are 1D such as ev_ratio...
-                    X_transf.set_coordset(
-                        x=Coord(
-                            None,
-                            labels=["#%d" % (i) for i in range(X_transf.shape[-1])],
-                            title="components",
-                        ),
-                    )
-                if self.typesingle == "targets":
-                    # occurs when the data are 1D such as PLSRegression intercept...
-                    if X.coordset[0].labels is not None:
-                        labels = X.coordset[0].labels
-                    else:
-                        labels = ["#%d" % (i + 1) for i in range(X.shape[-1])]
-                    X_transf.set_coordset(
-                        x=Coord(
-                            None,
-                            labels=labels,
-                            title="targets",
-                        ),
-                    )
-
-            # eventually restore masks
-            X_transf = obj._restore_masked_data(X_transf, axis=axis)
-            out.append(X_transf.squeeze())
-
-        if len(out) == 1:
-            return out[0]
-        else:
-            return tuple(out)
-
-
-def _wrap_ndarray_output_to_nddataset(
-    method=None,
-    meta_from="_X",
-    units="keep",
-    title="keep",
-    typex=None,
-    typey=None,
-    typesingle=None,
-):
-    # wrap _set_output to allow for deferred calling
-    if method:
-        # case of the decorator without argument
-        out = _set_output(method)
-    else:
-        # and with argument
-        def wrapper(method):
-            return _set_output(
-                method,
-                meta_from=meta_from,
-                units=units,
-                title=title,
-                typex=typex,
-                typey=typey,
-                typesingle=typesingle,
-            )
-
-        out = wrapper
-    return out
-
-
-# ======================================================================================
 # Base class AnalysisConfigurable
 # ======================================================================================
-class AnalysisConfigurable(MetaConfigurable):
-
+class AnalysisConfigurable(BaseConfigurable):
     __doc__ = _docstring.dedent(
         r"""
     Abstract class to write analysis model estimators.
@@ -264,8 +40,7 @@ class AnalysisConfigurable(MetaConfigurable):
 
     Parameters
     ----------
-    log_level : any of [``"INFO"``\ , ``"DEBUG"``\ , ``"WARNING"``\ , ``"ERROR"``\ ], optional, default: ``"WARNING"``
-        The log level at startup.
+    %(BaseConfigurable.parameters.log_level)s
     warm_start : `bool`\ , optional, default: `False`
         When fitting repeatedly on the same dataset, but for multiple
         parameter values (such as to find the value maximizing performance),
@@ -284,21 +59,7 @@ class AnalysisConfigurable(MetaConfigurable):
     # Runtime Parameters
     # ----------------------------------------------------------------------------------
     _fitted = tr.Bool(False, help="False if the model was not yet fitted")
-    _masked_rc = tr.Tuple(allow_none=True, help="List of masked rows and columns")
-    _X = NDDatasetType(allow_none=True, help="Data to fit a model")
-    _X_mask = Array(allow_none=True, help="Mask information of the input X data")
-    _X_preprocessed = Array(help="Preprocessed inital input X data")
-    _X_shape = tr.Tuple(
-        help="Original shape of the input X data, " "before any transformation"
-    )
-    _X_coordset = tr.Instance(CoordSet, allow_none=True)
-    _is_dataset = tr.Bool(help="True if the input X data is a NDDataset")
     _outfit = tr.Any(help="the output of the _fit method - generally a tuple")
-    _output_type = tr.Enum(
-        ["NDDataset", "ndarray"],
-        default_value="NDDataset",
-        help="Whether the output is a NDDataset or a ndarray",
-    )
 
     # ----------------------------------------------------------------------------------
     # Configuration parameters (mostly defined in subclass
@@ -319,195 +80,14 @@ class AnalysisConfigurable(MetaConfigurable):
         **kwargs,
     ):
         """ """
-        # An empty __doc__ is placed here, else Configurable.__doc__
-        # will appear when there is no __init___.doc in subclass
+        self._warm_start = warm_start
 
-        # Reset default configuration if not warm_start
-        reset = not warm_start
+        super().__init__(log_level=log_level, **kwargs)
 
-        # Call the super class (MetaConfigurable) for initialisation
-        super().__init__(parent=app, reset=reset)
-
-        # Set log_level of the console report (accessible using the log property)
-        set_loglevel(log_level)
-
-        # Initial configuration
-        # ---------------------
-        # Reset all config parameters to default, if not warm_start
-        defaults = self.parameters(default=True)
-        configkw = {} if warm_start else defaults
-
-        # Eventually take parameters from kwargs
-        configkw.update(kwargs)
-
-        # Now update all configuration parameters
-        # if an item k is not in the config parameters, an error is raised.
-        for k, v in configkw.items():
-            if hasattr(self, k) and k in defaults:
-                if getattr(self, k) != v:
-                    setattr(self, k, v)
-            else:
-                raise KeyError(
-                    f"'{k}' is not a valid configuration parameters. "
-                    f"Use the method `parameters()` to check the current "
-                    f"allowed parameters and their current value."
-                )
-
-        # If warm start we can use the previous fit as starting profiles.
-        # so the flag _fitted is not set.
         if not warm_start:
             # We should not be able to use any methods requiring fit results
             # until the fit method has been executed
             self._fitted = False
-
-    # ----------------------------------------------------------------------------------
-    # Private methods
-    # ----------------------------------------------------------------------------------
-    def _make_dataset(self, d):
-        # Transform an array-like object to NDDataset
-        # or a list of array-like to a list of NDQataset
-        if d is None:
-            return
-        if isinstance(d, (tuple, list)):
-            d = [self._make_dataset(item) for item in d]
-        elif not isinstance(d, NDDataset):
-            d = NDDataset(d, copy=True)
-        else:
-            d = d.copy()
-        return d
-
-    def _make_unsqueezed_dataset(self, d):
-        # add a dimension to 1D Dataset
-        if d.ndim == 1:
-            coordset = d.coordset
-            d._data = d._data[np.newaxis]
-            if np.any(d.mask):
-                d._mask = d._mask[np.newaxis]
-            d.dims = ["y", "x"]  # "y" is the new dimension
-            coordx = coordset[0] if coordset is not None else None
-            d.set_coordset(x=coordx, y=None)
-        return d
-
-    def _get_masked_rc(self, mask):
-        # Get the mask by row and columns.
-        # -------------------------------
-        # When a single element in the array is
-        # masked, the whole row and columns for this element is masked as well as the
-        # corresponding columns.
-        if np.any(mask):
-            masked_columns = np.all(mask, axis=-2)  # if mask.ndim == 2 else None
-            masked_rows = np.all(mask, axis=-1)
-        else:
-            masked_columns = np.zeros(self._X_shape[-1], dtype=bool)
-            masked_rows = np.zeros(self._X_shape[-2], dtype=bool)
-        return masked_rows, masked_columns
-
-    def _remove_masked_data(self, X):
-        # Retains only valid rows and columns
-        # -----------------------------------
-        # unfortunately, the implementation of linalg library
-        # doesn't support numpy masked arrays as input. So we will have to
-        # remove the masked values ourselves
-
-        # the following however assumes that entire rows or columns are masked,
-        # not only some individual data (if this is what you wanted, this
-        # will fail)
-        if not hasattr(X, "mask") or not np.any(X._mask):
-            return X
-
-        # remove masked rows and columns
-        masked_rows, masked_columns = self._get_masked_rc(X._mask)
-
-        Xc = X[:, ~masked_columns]
-        Xrc = Xc[~masked_rows]
-
-        # destroy the mask
-        Xrc._mask = NOMASK
-
-        # return the modified X dataset
-        return Xrc
-
-    def _restore_masked_data(self, D, axis=-1):
-        # by default, we restore columns, put axis=0 to restore rows instead
-        # Note that it is very important to use here the ma version of zeros
-        # array constructor or both if both axis should be restored
-        if not np.any(self._X_mask):
-            # return it inchanged as wa had no mask originally
-            return D
-
-        rowsize, colsize = self._X_shape
-        masked_rows, masked_columns = self._get_masked_rc(self._X_mask)
-
-        if D.ndim == 2:
-            # Put back masked columns in D
-            # ----------------------------
-            M, N = D.shape
-            if axis == "both":  # and D.shape[0] == rowsize:
-                if np.any(masked_columns) or np.any(masked_rows):
-                    Dtemp = np.ma.zeros((rowsize, colsize))  # note np.ma, not np.
-                    Dtemp[~self._X_mask] = D.data.flatten()
-                    Dtemp[self._X_mask] = MASKED
-                    D.data = Dtemp
-                    try:
-                        D.coordset[D.dims[-1]] = self._X_coordset[D.dims[-1]]
-                        D.coordset[D.dims[-2]] = self._X_coordset[D.dims[-2]]
-                    except TypeError:
-                        # probably no coordset
-                        pass
-            elif axis == -1 or axis == 1:
-                if np.any(masked_columns):
-                    Dtemp = np.ma.zeros((M, colsize))  # note np.ma, not np.
-                    Dtemp[:, ~masked_columns] = D
-                    Dtemp[:, masked_columns] = MASKED
-                    D.data = Dtemp
-                    try:
-                        D.coordset[D.dims[-1]] = self._X_coordset[D.dims[-1]]
-                    except TypeError:
-                        # probably no coordset
-                        pass
-
-            # Put back masked rows in D
-            # -------------------------
-            elif axis == -2 or axis == 0:
-                if np.any(masked_rows):
-                    Dtemp = np.ma.zeros((rowsize, N))
-                    Dtemp[~masked_rows] = D
-                    Dtemp[masked_rows] = MASKED
-                    D.data = Dtemp
-                    try:
-                        D.coordset[D.dims[-2]] = self._X_coordset[D.dims[-2]]
-                    except TypeError:
-                        # probably no coordset
-                        pass
-        elif D.ndim == 1:
-            # we assume here that the only case it happens is for array as explained
-            # variance so that we deal with masked rows
-            if np.any(masked_rows):
-                Dtemp = np.ma.zeros((rowsize,))  # note np.ma, not np.
-                Dtemp[~masked_rows] = D
-                Dtemp[masked_rows] = MASKED
-                D.data = Dtemp
-
-        elif D.ndim == 3:
-            # CASE of IRIS for instance
-
-            # Put back masked columns in D
-            # ----------------------------
-            J, M, N = D.shape
-            if axis == -1 or axis == 2:
-                if np.any(masked_columns):
-                    Dtemp = np.ma.zeros((J, M, colsize))  # note np.ma, not np.
-                    Dtemp[..., ~masked_columns] = D
-                    Dtemp[..., masked_columns] = MASKED
-                    D.data = Dtemp
-                    try:
-                        D.coordset[D.dims[-1]] = self._X_coordset[D.dims[-1]]
-                    except TypeError:
-                        # probably no coordset
-                        pass
-
-        # return the D array with restored masked data
-        return D
 
     # ----------------------------------------------------------------------------------
     # Private validation and default getter methods
@@ -515,32 +95,6 @@ class AnalysisConfigurable(MetaConfigurable):
     @tr.default("_X")
     def _X_default(self):
         raise NotFittedError
-
-    @tr.validate("_X")
-    def _X_validate(self, proposal):
-        # validation fired when self._X is assigned
-        X = proposal.value
-
-        # for the following we need X with two dimensions
-        # So let's generate the un-squeezed X
-        X = self._make_unsqueezed_dataset(X)
-
-        # as in fit methods we often use np.linalg library, we cannot handle directly
-        # masked data (so we remove them here and they will be restored at the end of
-        # the process during transform or inverse transform methods
-        # store the original shape as it will be eventually modified as welle- as the
-        # original coordset
-        self._X_shape = X.shape
-
-        # store the mask because it may be destroyed
-        self._X_mask = X._mask.copy()
-
-        # and the original coordset
-        self._X_coordset = copy(X._coordset)
-
-        # remove masked data and return modified dataset
-        X = self._remove_masked_data(X)
-        return X
 
     @property
     def _X_is_missing(self):
@@ -555,16 +109,6 @@ class AnalysisConfigurable(MetaConfigurable):
     # ----------------------------------------------------------------------------------
     # Private methods that should be, most of the time, overloaded in subclass
     # ----------------------------------------------------------------------------------
-    @tr.observe("_X")
-    def _preprocess_as_X_changed(self, change):
-        # to be optionally replaced by user defined function (with the same name)
-        X = change.new
-        # .... preprocessing as scaling, centering, ... must return a ndarray with
-        #  same shape a X.data
-
-        # Set a X.data by default
-        self._X_preprocessed = X.data
-
     def _fit(self, X, Y=None):  # pragma: no cover
         #  Intended to be replaced in the subclasses by user defined function
         #  (with the same name)
@@ -1525,7 +1069,6 @@ class CrossDecompositionAnalysis(DecompositionAnalysis):
 # Base class LinearRegressionAnalysis
 # ======================================================================================
 class LinearRegressionAnalysis(AnalysisConfigurable):
-
     # ----------------------------------------------------------------------------------
     # Configuration parameters (mostly defined in subclass
     # as they depend on the model estimator)
@@ -1559,7 +1102,6 @@ class LinearRegressionAnalysis(AnalysisConfigurable):
         warm_start=False,
         **kwargs,
     ):
-
         # call the super class for initialisation of the configuration parameters
         # to do before anything else!
         super().__init__(
@@ -1721,7 +1263,6 @@ class LinearRegressionAnalysis(AnalysisConfigurable):
             if self._is_dataset and self._Y.has_units and self._X.has_units:
                 A = A * self._Y.units / self._X.units
         elif self._is_dataset:
-
             unitsX = self._X.units if self._X.units is not None else 1.0
             unitsY = self._Y.units if self._Y.units is not None else 1.0
             if unitsX != 1 or unitsY != 1:
