+++ spectrochempy/analysis/baseline/__init__.py
@@ -1,0 +1,11 @@
+# # -*- coding: utf-8 -*-
+# # ======================================================================================
+# # Copyright (©) 2015-2023 LCS - Laboratoire Catalyse et Spectrochimie, Caen, France.
+# # CeCILL-B FREE SOFTWARE LICENSE AGREEMENT
+# # See full LICENSE agreement in the root directory.
+# # ======================================================================================
+# from spectrochempy.analysis.baseline import baseline as _baseline
+# from spectrochempy.analysis.baseline.baseline import *  # noqa: F401, F403
+# __all__ = _baseline.__all__
+# __configurables__ = _baseline.__configurables__
+# __dataset_methods__ = _baseline.__dataset_methods__
