+++ spectrochempy/analysis/baseline/utils.py
