+++ spectrochempy/analysis/decomposition/iris.py
@@ -18,7 +18,10 @@ import traitlets as tr
 from matplotlib import pyplot as plt
 from scipy import optimize
 
-from spectrochempy.analysis._base import DecompositionAnalysis, NotFittedError
+from spectrochempy.analysis._baseclass._analysisbase import (
+    DecompositionAnalysis,
+    NotFittedError,
+)
 from spectrochempy.core import info_, warning_
 from spectrochempy.core.dataset.coord import Coord
 from spectrochempy.core.dataset.coordset import CoordSet
