+++ spectrochempy/analysis/kinetics/kineticutilities.py
