+++ spectrochempy/analysis/peakfinding/peakfinding.py
