--- spectrochempy/analysis/preprocessing/__init__.py
@@ -1,22 +1,0 @@
-from spectrochempy.analysis.preprocessing.baseline import (
-    Baseline,
-    asls,
-    basc,
-    baseline,
-    detrend,
-    snip,
-)
-from spectrochempy.analysis.preprocessing.utils import lls, lls_inv
-
-__all__ = [
-    "Baseline",
-    "baseline",
-    "basc",
-    "detrend",
-    "asls",
-    "snip",
-    "lls",
-    "lls_inv",
-]
-__configurables__ = ["Baseline"]
-__dataset_methods__ = ["baseline", "basc", "detrend", "asls", "snip"]
