+++ spectrochempy/widgets/baselinecorrector.py
@@ -10,7 +10,7 @@ from functools import partial
 from IPython.display import display
 from ipywidgets import Layout, widgets
 
-from spectrochempy.analysis.preprocessing.baseline import Baseline
+from spectrochempy.analysis.baseline.baseline import Baseline
 from spectrochempy.core import info_, warning_
 from spectrochempy.core.dataset.nddataset import NDDataset
 from spectrochempy.core.plotters.multiplot import multiplot
